TO BE REVISITED: changed statistics recording
I've tried to change a couple of statistics to use the built-in recorders.
Unfortunately it's not obvious to me what were the original intentions,
so I'm not sure if this patch is correct or not. The idea is to call
emit() always on the module which is supposed to record the statistic. For
cell wide statistics this module is the MAC of the eNodeB, otherwise it's
the MAC of the Ue. I believe the old implementation was confusing, because
the emit() was called on a module which had no effect on the name of the
recorded statistics. The old recorders also use the MAC id to lookup the
recording state in maps. I think this is both inefficient and unnecessary.
Moreover, the macCellPacketLoss_ seems to be wrong to me, because the old
implementation recorded one value for each Ue? with the same statistics
name (the MAC of eNodeB)?

This patch is just a small part of this kind of refactoring. There are
quite a few other places in the code where this kind of statistics recording
happens.