Ensure HarqBufferRx only accesses valid source UEs
Each LteHarqBufferRx has a refernece to the source of the
current packet in the buffer. If the data is correct the
buffer will emit a signal for staticis *on* the source
macUe. If the source UE cModule left the simulation this
creats a SEGV. Changes ensure that each reference to
the removed macUe is nulled and emit calls will check
before accessing the macUe. The rest of the packet
processing is not changed. The removed UE will be
treated as a node wich is not reachable anymore.