simulation cars: ini file corrected, workaround for using HostAutoConfigurator of inet 4.2.2.
Includes workaround for two issues when using the HostAutoConfigurator:

* HostAutoConfigurator in inet 4.2.2 adds interface data in stage INITSTAGE_NETWORK_INTERFACE_CONFIGURATION,
but cellular LteNic interface is created in same stage by the inet framework. In the long term, it might
make sense to use INITSTAGE_NETWORK_CONFIGURATION for adding interface data in HostAutoConfigurator.
Until this is done, the LteNic must get to stage INITSTAGE_NETWORK_INTERFACE_CONFIGURATION before the
HostAutoConfigurator, therefore the structure of Node Ue was updated to list the LteNic before the ip layer.

* Furthermore, the HostAutoConfigurator calls setBroadcast(true) in setupNetworkLayer at
  INITSTAGE_NETWORK_CONFIGURATION == 4. For LteNic, which does not support broadcast, this needs to
  be corrected - othewise the node will try to send ARP requests. As a workaround, the IP2lte now
  calls setBroadcast(false) in INITSTAGE_STATIC_ROUTING.

Fingerprints had to be updated due to the modified order of initialization in Ue.